refactor: LottoBundle 테스트 추가
