+++ src/main/kotlin/domain/LottoBundle.kt
@@ -1,3 +1,5 @@
 package domain
 
-class LottoBundle(val lottos: List<Lotto>)
+class LottoBundle(lottoCount: Int, lottoGenerator: LottoGenerator) {
+    val lottos: List<Lotto> = List(lottoCount) { lottoGenerator.generate() }
+}
