--- src/test/kotlin/domain/LottoVendingMachineTest.kt
@@ -1,46 +1,0 @@
-package domain
-
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.ValueSource
-
-class LottoVendingMachineTest {
-    @ValueSource(ints = [1000, 5000, 10000])
-    @ParameterizedTest
-    fun `구입금액에 해당하는 로또 개수 계산`(amount: Int) {
-        // given
-        val money = Money(amount)
-
-        // when
-        val lottoCount: Int = LottoVendingMachine.getLottoCount(money)
-
-        // then
-        assertThat(lottoCount).isEqualTo(amount / LOTTO_PRICE)
-    }
-
-    @Test
-    fun `로또 개수만큼 발행`() {
-        // given
-        val lottoCount: Int = 5
-        val lotto = Lotto(
-            setOf(
-                LottoNumber.of(1),
-                LottoNumber.of(2),
-                LottoNumber.of(3),
-                LottoNumber.of(4),
-                LottoNumber.of(5),
-                LottoNumber.of(6)
-            )
-        )
-        // when
-        val lottoBundle = LottoVendingMachine.generateLottoBundle(lottoCount) { lotto }
-
-        // then
-        assertThat(lottoBundle.lottos.size).isEqualTo(lottoCount)
-    }
-
-    companion object {
-        private const val LOTTO_PRICE = 1000
-    }
-}
